requirements.txt
gunicorn
Flask==1.0.2
click==6.7
line-bot-sdk
azure-cognitiveservices-vision-computervision
imgur-python
Pillow